disable some of function and currently work in makefile
